feat: add attributes to Issue

+++ src/main/java/com/tuka/comiccharacters/service/IssueService.java
@@ -6,6 +6,7 @@ import com.tuka.comiccharacters.model.Issue;
 import com.tuka.comiccharacters.model.IssueCreator;
 import com.tuka.comiccharacters.model.Series;
 
+import java.math.BigDecimal;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
@@ -13,7 +14,7 @@ import java.util.Set;
 public class IssueService {
     private final IssueDaoImpl issueDao = new IssueDaoImpl();
 
-    public void addIssue(Series series, int issueNumber, List<IssueCreator> issueCreators, List<ComicCharacter> characters) {
+    public void addIssue(Series series, BigDecimal issueNumber, List<IssueCreator> issueCreators, List<ComicCharacter> characters) {
         Issue issue = new Issue(series, issueNumber);
 
         if (issueCreators != null && !issueCreators.isEmpty()) {
